+++ NEWS.md
@@ -107,6 +107,18 @@ unit = "s")
 
 12. `rbindlist` no longer errors when coercing complex vectors to character vectors, [#4202](https://github.com/Rdatatable/data.table/issues/4202). Thanks to @sritchie73 for reporting and the PR.
 
+13. Argument `keyby` of `[.data.table` can now be used as a _sorting_ flag when `by` argument is provided as well. In such case `keyby` is expected to be TRUE or FALSE, [#4307](https://github.com/Rdatatable/data.table/issues/4307).
+
+```r
+DT = data.table(a=2:1, b=3:2, d=4:3)
+
+DT[, .SD, by = "a"] ## same as the next one
+DT[, .SD, by = "a", keyby = FALSE]
+
+DT[, .SD, keyby = "a"] ## same as the next one
+DT[, .SD, by = "a", keyby = TRUE]
+```
+
 ## NOTES
 
 0. Retrospective license change permission was sought from and granted by 4 contributors who were missed in [PR#2456](https://github.com/Rdatatable/data.table/pull/2456), [#4140](https://github.com/Rdatatable/data.table/pull/4140). We had used [GitHub's contributor page](https://github.com/Rdatatable/data.table/graphs/contributors) which omits 3 of these due to invalid email addresses, unlike GitLab's contributor page which includes the ids. The 4th omission was a PR to a script which should not have been excluded; a script is code too. We are sorry these contributors were not properly credited before. They have now been added to the contributors list as displayed on CRAN. All the contributors of code to data.table hold its copyright jointly; your contributions belong to you. You contributed to data.table when it had a particular license at that time, and you contributed on that basis. This is why in the last license change, all contributors of code were consulted and each had a veto.